+++ src/main/kotlin/com/terraformation/backend/search/SearchService.kt
@@ -3,13 +3,8 @@ package com.terraformation.backend.search
 import com.terraformation.backend.log.perClassLogger
 import com.terraformation.backend.search.field.SearchField
 import jakarta.inject.Named
-import org.jooq.Condition
 import org.jooq.DSLContext
-import org.jooq.Record
-import org.jooq.Select
-import org.jooq.SelectJoinStep
 import org.jooq.conf.ParamType
-import org.jooq.impl.DSL
 
 /**
  * Searches the database based on user-supplied search criteria. Since the user is allowed to
@@ -25,36 +20,6 @@ import org.jooq.impl.DSL
 @Named
 class SearchService(private val dslContext: DSLContext) {
   private val log = perClassLogger()
-
-  /** Returns a condition that filters search results based on a list of criteria. */
-  private fun filterResults(
-      rootPrefix: SearchFieldPrefix,
-      criteria: Map<SearchFieldPrefix, SearchNode>
-  ): Condition {
-    // Filter out results the user doesn't have the ability to see. NestedQueryBuilder will include
-    // the visibility check on the root table, but not on parent tables.
-    val rootTable = rootPrefix.root
-    val rootCriterion = criteria[rootPrefix] ?: return DSL.trueCondition()
-    val conditions =
-        listOfNotNull(
-            rootCriterion.toCondition(),
-            rootTable.inheritsVisibilityFrom?.let { conditionForVisibility(it) })
-
-    val primaryKey = rootTable.primaryKey
-
-    val subquery =
-        joinWithSecondaryTables(
-                DSL.select(primaryKey).from(rootTable.fromTable), rootPrefix, rootCriterion)
-            .where(conditions)
-
-    // Ideally we'd preserve the type of the primary key column returned by the subquery, but that
-    // would require adding the primary key class as a type parameter in tons of places throughout
-    // the search code. (Try it if you're bored; you'll see it quickly spirals out of control!)
-    // The tiny amount of extra type safety we'd gain isn't worth the amount of boilerplate it'd
-    // require, especially since the primary key type isn't known at compile time anyway.
-    @Suppress("UNCHECKED_CAST")
-    return primaryKey.`in`(subquery as Select<Nothing>)
-  }
 
   /**
    * Queries the values of a list of fields that match a list of filter criteria.
@@ -121,9 +86,9 @@ class SearchService(private val dslContext: DSLContext) {
       sortOrder: List<SearchSortField> = emptyList(),
   ): NestedQueryBuilder {
     val queryBuilder = NestedQueryBuilder(dslContext, rootPrefix)
-    queryBuilder.addSelectFields(fields)
+    queryBuilder.addSelectFields(fields, criteria)
     queryBuilder.addSortFields(sortOrder)
-    queryBuilder.addCondition(filterResults(rootPrefix, criteria))
+    queryBuilder.addCondition(queryBuilder.filterResults(rootPrefix, criteria[rootPrefix]))
 
     return queryBuilder
   }
@@ -215,73 +180,5 @@ class SearchService(private val dslContext: DSLContext) {
     // SearchField.computeValue() can introduce duplicates that the query's SELECT DISTINCT has no
     // way of filtering out.
     return searchResults.map { it?.get(fieldPathName)?.toString() }.distinct()
-  }
-
-  /**
-   * Joins a query with any additional tables that are needed in order to determine which results
-   * the user has the ability to see. This is needed when the query's root prefix points at a table
-   * that doesn't include enough information to do visibility filtering.
-   *
-   * This can potentially join with multiple additional tables if the required information is more
-   * than one hop away in the graph of search tables.
-   *
-   * The resulting query will include all the tables that will be referenced by
-   * [conditionForVisibility].
-   *
-   * @param referencedTables Which tables are already referenced in the query. This method will not
-   *   join with these tables. Should not include tables that are only referenced in subqueries.
-   */
-  private fun <T : Record> joinForVisibility(
-      query: SelectJoinStep<T>,
-      referencedTables: Set<SearchTable>,
-      searchTable: SearchTable
-  ): SelectJoinStep<T> {
-    val inheritsVisibilityFrom = searchTable.inheritsVisibilityFrom ?: return query
-
-    return if (inheritsVisibilityFrom in referencedTables) {
-      // We've already joined with the next table in the chain, so no need to do it again. But we
-      // might still need to join with additional tables beyond the next one.
-      joinForVisibility(query, referencedTables, inheritsVisibilityFrom)
-    } else {
-      // The query doesn't already include the table we need to join with from this one in order to
-      // evaluate visibility; join with it and then see if there are additional tables that also
-      joinForVisibility(
-          searchTable.joinForVisibility(query),
-          referencedTables + inheritsVisibilityFrom,
-          inheritsVisibilityFrom)
-    }
-  }
-
-  /**
-   * Returns a condition that checks whether the user is able to view a particular search result.
-   *
-   * The condition can refer to columns in any tables that are added by [joinForVisibility].
-   */
-  private fun conditionForVisibility(searchTable: SearchTable): Condition? {
-    return searchTable.conditionForVisibility()
-        ?: searchTable.inheritsVisibilityFrom?.let { conditionForVisibility(it) }
-  }
-
-  /**
-   * Adds JOIN clauses to a query to join the root table with any other tables referenced by a list
-   * of [SearchField] s. This is not used for nested fields.
-   *
-   * This handles indirect references; if a field is in a table that is two foreign-key hops away
-   * from `accession`, the intermediate table is included here too.
-   */
-  private fun <T : Record> joinWithSecondaryTables(
-      selectFrom: SelectJoinStep<T>,
-      rootPrefix: SearchFieldPrefix,
-      criteria: SearchNode,
-  ): SelectJoinStep<T> {
-    val referencedSublists = criteria.referencedSublists().distinctBy { it.searchTable }
-    val referencedTables = referencedSublists.map { it.searchTable }.toSet()
-
-    val joinedQuery =
-        referencedSublists.fold(selectFrom) { query, sublist ->
-          query.leftJoin(sublist.searchTable.fromTable).on(sublist.conditionForMultiset)
-        }
-
-    return joinForVisibility(joinedQuery, referencedTables, rootPrefix.root)
   }
 }
